+++ app/src/androidTest/java/com/github/steroidteam/todolist/ListSelectionActivityTest.java
@@ -1,13 +1,13 @@
 package com.github.steroidteam.todolist;
 
-import android.content.Intent;
-import androidx.test.core.app.ActivityScenario;
-import androidx.test.core.app.ApplicationProvider;
-import androidx.test.espresso.Espresso;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.intent.matcher.IntentMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import com.github.steroidteam.todolist.view.ItemViewActivity;
+import com.github.steroidteam.todolist.view.ListSelectionActivity;
+import com.github.steroidteam.todolist.view.NoteSelectionActivity;
 
 import org.hamcrest.Matchers;
 import org.junit.Rule;
@@ -15,9 +15,6 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.closeSoftKeyboard;
-import static androidx.test.espresso.action.ViewActions.typeText;
-import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static org.hamcrest.Matchers.anything;
