--- app/src/main/java/com/github/steroidteam/todolist/ListSelectionActivity.java
@@ -1,43 +1,0 @@
-package com.github.steroidteam.todolist;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.ActionBar;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-
-import com.firebase.ui.auth.AuthUI;
-import com.google.android.gms.auth.api.Auth;
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-
-public class ListSelectionActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_list_selection);
-        //Toolbar toolbar = findViewById(R.id.list_selection_toolbar);
-        //setSupportActionBar(toolbar);
-        //ActionBar actionBar = getSupportActionBar();
-        //actionBar.setDisplayHomeAsUpEnabled(true);
-    }
-
-    public void logOut(View view) {
-        Activity thisActivity = this;
-        AuthUI.getInstance().signOut(this).addOnCompleteListener(task -> thisActivity.finish());
-    }
-
-    public void openList(View view){
-        Intent itemViewActivity = new Intent(ListSelectionActivity.this, ItemViewActivity.class);
-        startActivity(itemViewActivity);
-    }
-
-    public void openNotes(View view) {
-        Intent noteSelectionActivity = new Intent(ListSelectionActivity.this, NoteSelectionActivity.class);
-        startActivity(noteSelectionActivity);
-    }
-}
