+++ app/src/main/java/com/github/steroidteam/todolist/database/Database.java
@@ -1,7 +1,7 @@
 package com.github.steroidteam.todolist.database;
 
-import com.github.steroidteam.todolist.todo.Task;
-import com.github.steroidteam.todolist.todo.TodoList;
+import com.github.steroidteam.todolist.model.todo.Task;
+import com.github.steroidteam.todolist.model.todo.TodoList;
 
 import java.util.UUID;
