+++ app/src/main/java/com/github/steroidteam/todolist/database/FirebaseDatabase.java
@@ -3,11 +3,9 @@ package com.github.steroidteam.todolist.database;
 import androidx.annotation.NonNull;
 
 import com.github.steroidteam.todolist.filestorage.FirebaseFileStorageService;
-import com.github.steroidteam.todolist.todo.Task;
-import com.github.steroidteam.todolist.todo.TodoList;
+import com.github.steroidteam.todolist.model.todo.Task;
+import com.github.steroidteam.todolist.model.todo.TodoList;
 import com.github.steroidteam.todolist.util.JSONSerializer;
-import com.google.firebase.auth.FirebaseUser;
-import com.google.firebase.storage.FirebaseStorage;
 
 import java.nio.charset.StandardCharsets;
 import java.util.Objects;
