+++ app/src/main/java/com/github/steroidteam/todolist/model/TodoRepository.java
@@ -1,10 +1,11 @@
-package com.github.steroidteam.todolist.database;
+package com.github.steroidteam.todolist.model;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 
-import com.github.steroidteam.todolist.todo.Task;
-import com.github.steroidteam.todolist.todo.TodoList;
+import com.github.steroidteam.todolist.database.VolatileDatabase;
+import com.github.steroidteam.todolist.model.todo.Task;
+import com.github.steroidteam.todolist.model.todo.TodoList;
 
 import java.util.UUID;
