+++ app/src/main/java/com/github/steroidteam/todolist/model/notes/Note.java
@@ -1,4 +1,4 @@
-package com.github.steroidteam.todolist.notes;
+package com.github.steroidteam.todolist.model.notes;
 
 import androidx.annotation.NonNull;
