+++ app/src/main/java/com/github/steroidteam/todolist/model/todo/Task.java
@@ -1,4 +1,4 @@
-package com.github.steroidteam.todolist.todo;
+package com.github.steroidteam.todolist.model.todo;
 
 import java.util.Objects;
