+++ app/src/main/java/com/github/steroidteam/todolist/model/todo/TodoList.java
@@ -1,4 +1,4 @@
-package com.github.steroidteam.todolist.todo;
+package com.github.steroidteam.todolist.model.todo;
 
 import androidx.annotation.Nullable;
