+++ app/src/main/java/com/github/steroidteam/todolist/util/JSONSerializer.java
@@ -2,7 +2,7 @@ package com.github.steroidteam.todolist.util;
 
 import androidx.annotation.NonNull;
 
-import com.github.steroidteam.todolist.todo.TodoList;
+import com.github.steroidteam.todolist.model.todo.TodoList;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
