+++ app/src/main/java/com/github/steroidteam/todolist/view/adapter/TodoAdapter.java
@@ -1,4 +1,4 @@
-package com.github.steroidteam.todolist.util;
+package com.github.steroidteam.todolist.view.adapter;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -12,8 +12,8 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.github.steroidteam.todolist.R;
-import com.github.steroidteam.todolist.todo.Task;
-import com.github.steroidteam.todolist.todo.TodoList;
+import com.github.steroidteam.todolist.model.todo.Task;
+import com.github.steroidteam.todolist.model.todo.TodoList;
 
 public class TodoAdapter extends RecyclerView.Adapter<TodoAdapter.TaskHolder> {
