+++ app/src/main/res/layout/activity_item_view.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ItemViewActivity">
+    tools:context=".view.ItemViewActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/activity_itemview_toolbar"
