+++ app/src/main/res/layout/activity_list_selection.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ListSelectionActivity">
+    tools:context=".view.ListSelectionActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/list_selection_toolbar2"
