+++ app/src/main/res/layout/activity_note_display.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NoteDisplayActivity">
+    tools:context=".view.NoteDisplayActivity">
 
     <androidx.appcompat.widget.Toolbar
         android:id="@+id/activity_notedisplay_toolbar"
