+++ app/src/test/java/com/github/steroidteam/todolist/model/todo/TaskTest.java
@@ -1,6 +1,4 @@
-package com.github.steroidteam.todolist.todo;
-
-import com.github.steroidteam.todolist.todo.Task;
+package com.github.steroidteam.todolist.model.todo;
 
 import org.junit.Test;
